# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SlideBanai API",
    "description": "AI-Powered Presentation Platform with Analytics & Collaboration",
    "version": "1.0.0",
    "contact": {
      "name": "SlideBanai Support",
      "url": "https://slidebanai.com/support",
      "email": "support@slidebanai.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.slidebanai.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token"
      }
    },
    "schemas": {
...
      "LearningDashboard": {
        "type": "object",
        "properties": {
          "user_id": {"type": "integer"},
          "total_exercises_completed": {"type": "integer"},
          "current_streak": {"type": "integer"},
          "total_learning_hours": {"type": "number"},
          "skill_level": {"type": "string", "enum": ["beginner", "intermediate", "advanced", "expert"]},
          "next_exercise": {"$ref": "#/components/schemas/Exercise"},
          "recent_milestones": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/Milestone"}
          }
        }
      },
      "LearningProfile": {
        "type": "object",
        "properties": {
          "user_id": {"type": "integer"},
          "learning_style": {"type": "string"},
          "preferred_difficulty": {"type": "string"},
          "completed_courses": {"type": "array", "items": {"type": "string"}},
          "skill_assessments": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/SkillProgress"}
          },
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "Exercise": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "title": {"type": "string"},
          "description": {"type": "string"},
          "difficulty": {"type": "string", "enum": ["easy", "medium", "hard"]},
          "estimated_time_minutes": {"type": "integer"},
          "topic": {"type": "string"},
          "is_completed": {"type": "boolean"},
          "completion_percentage": {"type": "number"},
          "estimated_points": {"type": "integer"}
        }
      },
      "ExerciseCompletion": {
        "type": "object",
        "properties": {
          "exercise_id": {"type": "integer"},
          "score": {"type": "number"},
          "time_spent_seconds": {"type": "integer"},
          "points_earned": {"type": "integer"},
          "feedback": {"type": "string"},
          "completed_at": {"type": "string", "format": "date-time"}
        }
      },
      "SkillProgress": {
        "type": "object",
        "properties": {
          "skill_id": {"type": "integer"},
          "skill_name": {"type": "string"},
          "proficiency_level": {"type": "string", "enum": ["novice", "beginner", "intermediate", "advanced", "expert"]},
          "progress_percentage": {"type": "number"},
          "exercises_completed": {"type": "integer"},
          "total_exercises": {"type": "integer"},
          "last_practiced": {"type": "string", "format": "date-time"}
        }
      },
      "Milestone": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "title": {"type": "string"},
          "description": {"type": "string"},
          "is_completed": {"type": "boolean"},
          "completion_date": {"type": "string", "format": "date-time"},
          "required_exercises": {"type": "integer"},
          "exercises_completed": {"type": "integer"}
        }
      },
      "AIInsights": {
        "type": "object",
        "properties": {
          "user_id": {"type": "integer"},
          "strengths": {
            "type": "array",
            "items": {"type": "string"}
          },
          "weaknesses": {
            "type": "array",
            "items": {"type": "string"}
          },
          "recommendations": {
            "type": "array",
            "items": {"type": "string"}
          },
          "learning_pace": {"type": "string"},
          "estimated_time_to_mastery": {"type": "string"},
          "generated_at": {"type": "string", "format": "date-time"}
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/analytics/presentation/{presentationId}": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get presentation analytics",
        "operationId": "getPresentationAnalytics",
        "parameters": [
          {
            "name": "presentationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Presentation analytics"
          }
        }
      }
    },
...
    "/api/adaptive-learning/dashboard": {
      "get": {
        "tags": ["Adaptive Learning"],
        "summary": "Get learning dashboard",
        "operationId": "getLearningDashboard",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Learning dashboard retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "data": {
                      "$ref": "#/components/schemas/LearningDashboard"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/adaptive-learning/profile": {
      "get": {
        "tags": ["Adaptive Learning"],
        "summary": "Get learning profile",
        "operationId": "getLearningProfile",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Learning profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "data": {
                      "$ref": "#/components/schemas/LearningProfile"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/adaptive-learning/exercises": {
      "get": {
        "tags": ["Adaptive Learning"],
        "summary": "Get personalized exercises",
        "operationId": "getPersonalizedExercises",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "includeCompleted",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Include completed exercises in response"
          }
        ],
        "responses": {
          "200": {
            "description": "Personalized exercises retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Exercise"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/adaptive-learning/exercises/{exerciseId}/complete": {
      "post": {
        "tags": ["Adaptive Learning"],
        "summary": "Complete an exercise",
        "operationId": "completeExercise",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "exerciseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "score": {
                    "type": "number",
                    "description": "User's score on the exercise"
                  },
                  "time_spent_seconds": {
                    "type": "integer",
                    "description": "Time spent on exercise in seconds"
                  },
                  "feedback": {
                    "type": "string",
                    "description": "Optional feedback on performance"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exercise completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "data": {
                      "$ref": "#/components/schemas/ExerciseCompletion"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Exercise not found"
          }
        }
      }
    },
    "/api/adaptive-learning/skills": {
      "get": {
        "tags": ["Adaptive Learning"],
        "summary": "Get skill progress",
        "operationId": "getSkillProgress",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Skill progress retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SkillProgress"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/adaptive-learning/milestones": {
      "get": {
        "tags": ["Adaptive Learning"],
        "summary": "Get learning milestones",
        "operationId": "getLearningMilestones",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Learning milestones retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Milestone"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/adaptive-learning/insights": {
      "get": {
        "tags": ["Adaptive Learning"],
        "summary": "Get AI insights",
        "operationId": "getInsights",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "AI insights retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "data": {
                      "$ref": "#/components/schemas/AIInsights"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
}